requirements.txt
pandas>=2.1.0
great-expectations>=0.18.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
